unit tests: Move test setup code to ironic_test.go
There's no need to keep the code to set up a test provisioner in
factory.go since it is only used in unit tests, so move it to
ironic_test.go.
