+++ pkg/provisioner/ironic/ironic_test.go
@@ -11,6 +11,7 @@ import (
 	metal3v1alpha1 "github.com/metal3-io/baremetal-operator/apis/metal3.io/v1alpha1"
 	"github.com/metal3-io/baremetal-operator/pkg/bmc"
 	"github.com/metal3-io/baremetal-operator/pkg/provisioner"
+	"github.com/metal3-io/baremetal-operator/pkg/provisioner/ironic/clients"
 
 	// We don't use this package directly here, but need it imported
 	// so it registers its test fixture with the other BMC access
@@ -20,6 +21,40 @@ import (
 
 func init() {
 	logf.SetLogger(logz.New(logz.UseDevMode(true)))
+}
+
+func newTestProvisionerFactory() ironicProvisionerFactory {
+	return ironicProvisionerFactory{
+		log: logf.Log,
+		config: ironicConfig{
+			deployKernelURL:  "http://deploy.test/ipa.kernel",
+			deployRamdiskURL: "http://deploy.test/ipa.initramfs",
+			deployISOURL:     "http://deploy.test/ipa.iso",
+			maxBusyHosts:     20,
+		},
+	}
+}
+
+// A private function to construct an ironicProvisioner (rather than a
+// Provisioner interface) in a consistent way for tests.
+func newProvisionerWithSettings(host metal3v1alpha1.BareMetalHost, bmcCreds bmc.Credentials, publisher provisioner.EventPublisher, ironicURL string, ironicAuthSettings clients.AuthConfig, inspectorURL string, inspectorAuthSettings clients.AuthConfig) (*ironicProvisioner, error) {
+	hostData := provisioner.BuildHostData(host, bmcCreds)
+
+	tlsConf := clients.TLSConfig{}
+	clientIronic, err := clients.IronicClient(ironicURL, ironicAuthSettings, tlsConf)
+	if err != nil {
+		return nil, err
+	}
+
+	clientInspector, err := clients.InspectorClient(inspectorURL, inspectorAuthSettings, tlsConf)
+	if err != nil {
+		return nil, err
+	}
+
+	factory := newTestProvisionerFactory()
+	factory.clientIronic = clientIronic
+	factory.clientInspector = clientInspector
+	return factory.ironicProvisioner(hostData, publisher)
 }
 
 func makeHost() metal3v1alpha1.BareMetalHost {
